backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.10.6
pypdf2==3.0.1
chromadb==0.4.22
networkx==3.2.1
pyvis==0.3.2
openai>=1.12.0
python-multipart==0.0.6
python-dotenv==1.0.0
